refactor: push_payload - model 이동
